+++ applications/cluster-matrix/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "cluster-matrix"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+graphics-common = { path = "../../graphics-common" }
+cluster-core = { path = "../../cluster-core" }
+embedded-graphics = { workspace = true }
+libm = { workspace = true }
+heapless = { workspace = true }
+
+[features]
+default = []
+std = []
