+++ cluster-core/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "cluster-core"
+version = "0.1.0"
+edition = "2021"
+
+[features]
+default = []
+std = []
+
+[dependencies]
+graphics-common = { path = "../graphics-common" }
+embedded-graphics-core = { workspace = true }
+embedded-graphics = { workspace = true }
+heapless = { workspace = true }
+libm = { workspace = true }
