+++ cluster-core/src/lib.rs
@@ -1,0 +1,8 @@
+#![no_std]
+
+#[cfg(feature = "std")]
+extern crate std;
+
+pub mod constants;
+pub mod shared;
+pub mod visualization;
