+++ cluster-core/src/shared.rs
