+++ cluster-core/src/shared/types.rs
