+++ cluster-core/src/visualization/display.rs
