+++ cluster-core/src/visualization/seats.rs
