--- common/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "common"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-embedded-graphics = { workspace = true }
-libm = { workspace = true }
-heapless = { workspace = true }
-
-[features]
-default = []
-std = []
