--- common/src/lib.rs
@@ -1,10 +1,0 @@
-#![no_std]
-
-#[cfg(feature = "std")]
-extern crate std;
-
-pub mod animations;
-pub mod constants;
-pub mod shared;
-pub mod utilities;
-pub mod visualization;
