+++ graphics-common/src/animations.rs
