+++ graphics-common/src/animations/arrow.rs
