+++ graphics-common/src/animations/fortytwo.rs
