+++ graphics-common/src/lib.rs
@@ -1,0 +1,7 @@
+#![no_std]
+
+#[cfg(feature = "std")]
+extern crate std;
+
+pub mod animations;
+pub mod utilities;
