+++ graphics-common/src/utilities.rs
