+++ hardware-tests/basic-panel/rust-toolchain.toml
